requirements.txt
# Django core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0
django-encrypted-model-fields==0.6.5

# Authentication
PyJWT==2.8.0
cryptography==41.0.7
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Google APIs
google-api-python-client==2.111.0

# Celery
celery==5.3.4
redis==5.0.1
django-celery-beat==2.8.1
django-celery-results==2.5.1

# AI/ML
openai==1.6.1

# Web scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.4

# Utils
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0

# Development
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
black==23.12.1
flake8==6.1.0
isort==5.13.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.39.1